# feat(bulk-import): improve performance of the `GET /imports` endpoint [RHIDP-3680]