--- a/plugins/bulk-import-backend/src/service/router.test.ts
+++ b/plugins/bulk-import-backend/src/service/router.test.ts
@@ -741,33 +741,12 @@ describe('createRouter', () => {
       mockedAuthorize.mockImplementation(allowAll);
 
       jest
-        .spyOn(GithubApiService.prototype, 'getRepositoriesFromIntegrations')
-        .mockResolvedValue({
-          repositories: [
-            {
-              name: 'A',
-              full_name: 'my-ent-org-1/A',
-              url: 'https://api.github.com/repos/my-ent-org-1/A',
-              html_url: 'https://github.com/my-ent-org-1/A',
-              default_branch: 'dev',
-            },
-            {
-              name: 'B',
-              full_name: 'my-ent-org-1/B',
-              url: 'https://api.github.com/repos/my-ent-org-1/B',
-              html_url: 'https://github.com/my-ent-org-1/B',
-              default_branch: 'main',
-            },
-            {
-              name: 'A',
-              full_name: 'my-ent-org-2/A',
-              url: 'https://api.github.com/repos/my-ent-org-2/A',
-              html_url: 'https://github.com/my-ent-org-2/A',
-              default_branch: 'dev',
-            },
-          ],
-          errors: [],
-        });
+        .spyOn(CatalogInfoGenerator.prototype, 'listCatalogUrlLocations')
+        .mockResolvedValue([
+          'https://github.com/my-ent-org-1/A/blob/dev/catalog-info.yaml',
+          'https://github.com/my-ent-org-1/B/blob/main/catalog-info.yaml',
+          'https://github.com/my-ent-org-2/A/blob/dev/catalog-info.yaml',
+        ]);
       jest
         .spyOn(GithubApiService.prototype, 'findImportOpenPr')
         .mockImplementation((_logger, input) => {
@@ -798,22 +777,14 @@ describe('createRouter', () => {
             input.repoUrl === 'https://github.com/my-ent-org-1/A',
           );
         });
-      jest
-        .spyOn(CatalogInfoGenerator.prototype, 'listCatalogUrlLocations')
-        .mockResolvedValue([
-          'https://github.com/my-ent-org-1/A/blob/dev/catalog-info.yaml',
-        ]);
 
       const response = await request(app).get('/imports');
       expect(response.status).toEqual(200);
       expect(response.body).toEqual([
         {
           approvalTool: 'GIT',
-          id: 'my-ent-org-1/A',
+          id: 'https://github.com/my-ent-org-1/A',
           repository: {
-            defaultBranch: 'dev',
-            errors: [],
-            id: 'my-ent-org-1/A',
             name: 'A',
             organization: 'my-ent-org-1',
             url: 'https://github.com/my-ent-org-1/A',
@@ -825,11 +796,8 @@ describe('createRouter', () => {
           errors: [
             'could not find out if there is an import PR open on this repo',
           ],
-          id: 'my-ent-org-1/B',
+          id: 'https://github.com/my-ent-org-1/B',
           repository: {
-            defaultBranch: 'main',
-            errors: [],
-            id: 'my-ent-org-1/B',
             name: 'B',
             organization: 'my-ent-org-1',
             url: 'https://github.com/my-ent-org-1/B',
@@ -838,7 +806,7 @@ describe('createRouter', () => {
         },
         {
           approvalTool: 'GIT',
-          id: 'my-ent-org-2/A',
+          id: 'https://github.com/my-ent-org-2/A',
           github: {
             pullRequest: {
               number: 987,
@@ -846,9 +814,6 @@ describe('createRouter', () => {
             },
           },
           repository: {
-            defaultBranch: 'dev',
-            errors: [],
-            id: 'my-ent-org-2/A',
             name: 'A',
             organization: 'my-ent-org-2',
             url: 'https://github.com/my-ent-org-2/A',