--- a/plugins/bulk-import-backend/src/service/router.ts
+++ b/plugins/bulk-import-backend/src/service/router.ts
@@ -135,6 +135,7 @@ export async function createRouter(
       q.checkImportStatus = stringToBoolean(q.checkImportStatus);
       const response = await findAllRepositories(
         logger,
+        config,
         githubApiService,
         catalogInfoGenerator,
         q.checkImportStatus,
@@ -165,6 +166,7 @@ export async function createRouter(
       q.checkImportStatus = stringToBoolean(q.checkImportStatus);
       const response = await findRepositoriesByOrganization(
         logger,
+        config,
         githubApiService,
         catalogInfoGenerator,
         c.request.params.organizationName?.toString(),
@@ -195,6 +197,7 @@ export async function createRouter(
       q.sizePerIntegration = stringToNumber(q.sizePerIntegration);
       const response = await findAllImports(
         logger,
+        config,
         githubApiService,
         catalogInfoGenerator,
         q.pagePerIntegration,
@@ -242,6 +245,7 @@ export async function createRouter(
       }
       const response = await findImportStatusByRepo(
         logger,
+        config,
         githubApiService,
         catalogInfoGenerator,
         q.repo,
@@ -263,6 +267,7 @@ export async function createRouter(
       }
       const response = await deleteImportByRepo(
         logger,
+        config,
         githubApiService,
         catalogInfoGenerator,
         q.repo,