**What does this PR do / why we need it:**

This PR improves the overall bulk-import user experience by improving the performance of the `GET /imports` endpoint.

It should also fix the issue where the list of imports returned would sometimes not be accurate (some repositories just imported would not show up in this list depending on the page and size requested).

Before this PR, listing imports would take several seconds (and even more depending on the number of repositories accessible from the GH integrations). Example below where it took **~35 seconds** to return a response:

```
$ time http -A bearer -a "${ADMIN_CURL_TOKEN}" GET http://localhost:7007/api/bulk-import/imports pagePerIntegration==1 sizePerIntegration==10000

[
    {
        "approvalTool": "GIT",
        "id": "https://github.com/lemra-org-ent-1/java-quarkus-starter",
        "lastUpdate": "2024-08-19T19:45:11Z",
        "repository": {
            "name": "java-quarkus-starter",
            "organization": "lemra-org-ent-1",
            "url": "https://github.com/lemra-org-ent-1/java-quarkus-starter"
        },
        "status": "ADDED"
    },
    {
        "approvalTool": "GIT",
        "github": {
            "pullRequest": {
                "number": 36,
                "url": "https://github.com/lemra-org-ent-2/animated-happiness/pull/36"
            }
        },
        "id": "https://github.com/lemra-org-ent-2/animated-happiness",
        "lastUpdate": "2024-08-21T07:38:28Z",
        "repository": {
            "name": "animated-happiness",
            "organization": "lemra-org-ent-2",
            "url": "https://github.com/lemra-org-ent-2/animated-happiness"
        },
        "status": "WAIT_PR_APPROVAL"
    }
]

http -A bearer -a "${ADMIN_CURL_TOKEN}" GET  pagePerIntegration==1   0.25s user 0.03s system 0% cpu 35.426 total
```

This PR improves this by no longer iterating from the repository list. As we can see in the output below, the same call above now takes **less than 2 seconds** to return a response.
Now the response time depends on the number of Location entities (treated as Import jobs), and not on the number of repositories accessible from the GH integrations configured.
```
$ time http -A bearer -a "${ADMIN_CURL_TOKEN}" GET http://localhost:7007/api/bulk-import/imports pagePerIntegration==1 sizePerIntegration==10000

[
    {
        "approvalTool": "GIT",
        "id": "https://github.com/lemra-org-ent-1/java-quarkus-starter",
        "lastUpdate": "2024-08-19T19:45:11Z",
        "repository": {
            "name": "java-quarkus-starter",
            "organization": "lemra-org-ent-1",
            "url": "https://github.com/lemra-org-ent-1/java-quarkus-starter"
        },
        "status": "ADDED"
    },
    {
        "approvalTool": "GIT",
        "github": {
            "pullRequest": {
                "number": 36,
                "url": "https://github.com/lemra-org-ent-2/animated-happiness/pull/36"
            }
        },
        "id": "https://github.com/lemra-org-ent-2/animated-happiness",
        "lastUpdate": "2024-08-21T07:38:28Z",
        "repository": {
            "name": "animated-happiness",
            "organization": "lemra-org-ent-2",
            "url": "https://github.com/lemra-org-ent-2/animated-happiness"
        },
        "status": "WAIT_PR_APPROVAL"
    }
]

http -A bearer -a "${ADMIN_CURL_TOKEN}" GET  pagePerIntegration==1   0.26s user 0.03s system 24% cpu 1.193 total
```

**Which issue(s) this PR fixes:**

Fixes https://issues.redhat.com/browse/RHIDP-3680

**PR acceptance criteria:**

- [ ] Unit tests

- [ ] Integration tests

- [ ] Documentation 

**How to test changes / Special notes to the reviewer:**

/cc @debsmita1 @ciiay @jerolimov